makewn.py: Allow scraping merged pull requests from GitHub generated release notes.
We've now exceeded the number of closed pull requests that can be
retrieved via the REST API.  The only way to filter pull requests
merged between a tag and a branch tip on the server side is by drafting
a release and generating release notes.  This requires more manual
steps, but there doesn't seem to be a workable alternative.